requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysql-connector==2.1.6
mysql-connector-python==8.0.13
pkg-resources==0.0.0
protobuf==3.6.1
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
